+++ src/test/java/de/alpharogroup/comparators/pairs/KeyValuesPairKeyComparatorTest.java
@@ -32,7 +32,7 @@ import java.util.List;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.Test;
 
-import de.alpharogroup.collections.list.ListExtensions;
+import de.alpharogroup.collections.list.ListFactory;
 import de.alpharogroup.collections.pairs.KeyValuesPair;
 import de.alpharogroup.test.objects.Employee;
 import de.alpharogroup.test.objects.Person;
@@ -57,14 +57,14 @@ public class KeyValuesPairKeyComparatorTest
 	{
 		final Person person = Person.builder().name("John").married(Boolean.FALSE).build();
 		final Person person2 = Person.builder().name("Anton").married(Boolean.FALSE).build();
-		final List<Employee> employees = ListExtensions.newArrayList();
+		final List<Employee> employees = ListFactory.newArrayList();
 		employees.add(Employee.builder().id("10").build());
 		employees.add(Employee.builder().id("20").build());
 		final Employee employee = Employee.builder().id("20").build();
 		final KeyValuesPair<Person, Employee> o1 = KeyValuesPair.<Person, Employee> builder()
 			.key(person).values(employees).value(employee).build();
 
-		final List<Employee> employees2 = ListExtensions.newArrayList();
+		final List<Employee> employees2 = ListFactory.newArrayList();
 		employees2.add(Employee.builder().id("30").build());
 		employees2.add(Employee.builder().id("40").build());
 		final Employee employee2 = Employee.builder().id("30").build();